MFC r337508:
Terminate filter_create_ext() args with NULL, not 0.

filter_create_ext() is documented to take a NULL terminated set of
arguments.  0 is promoted to an int so this would fail on 64-bit
systems if the value was not passed in a register.  On all currently
supported 64-bit architectures it is.

Obtained from:	CheriBSD
Sponsored by:	DARPA, AFRL


git-svn-id: https://svn.freebsd.org/base/stable/11@337766 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f