MFC r337559: Makefile.inc1: Add libl to -legacy as well
libl is needed for config(8), which is a bootstrap-tool. It is possible to
build a system WITHOUT_TOOLCHAIN to exclude lex and thus, libl. We still
need to support building from this kind of host, though.

While here, group the config(8) dependencies together and add a small
explanation. These can likely both be scoped more clearly, but this will
need some further investigation.


git-svn-id: https://svn.freebsd.org/base/stable/11@337946 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f