MFC r336184: net80211: Fix ifdetach w/o ifattach, small whitespace cleanup
As the comment says, ifdetach might be called during the course of driver
detach if initialization failed. This shouldn't be a total failure, though,
we just have nothing to do there.

This has been modified slightly from Augustin's original commit to move the
bail-out slightly earlier since the ic wouldn't have been added to the
ic list in the first place, and a comment has been added describing when
this might be an issue.


git-svn-id: https://svn.freebsd.org/base/stable/11@337950 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f