arm64: allwinner: Add aw_syscon driver to GENERIC
Recent DTS use the syscon for the emac controller.
We support this but since U-Boot is still using old DTS it was never
needed for us to add this support, but this is a problem when using upstream
recent DTS and will be when U-Boot will catch up.

While here add a new compatible to the aw_syscon driver as Linux changed it ...


git-svn-id: https://svn.freebsd.org/base/head@338070 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f