MFC r312770 and r337854:
After the in_control() changes in r257692, an existing address is
(intentionally) deleted first and then completely added again (so all the
events, announces and hooks are given a chance to run).

This cause an issue with CARP where the existing CARP data structure is
removed together with the last address for a given VHID, which will cause
a subsequent fail when the address is later re-added.

This change fixes this issue by adding a new flag to keep the CARP data
structure when an address is not being removed.

There was an additional issue with IPv6 CARP addresses, where the CARP data
structure would never be removed after a change and lead to VHIDs which
cannot be destroyed.

PR:		229384
Sponsored by:	Rubicon Communications, LLC (Netgate)


git-svn-id: https://svn.freebsd.org/base/stable/11@338081 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f