MFC r338047:
The bucket index is subtracted by one at lines 2304 and 2314.  When 0 it
becomes -1, except these are unsigned integers, so they become very large
numbers. Thus are always larger than the maximum bucket; the hash table
insertion fails causing NAT to fail.

This commit ensures that if the index is already zero it is not reduced
prior to insertion into the hash table.

PR:		208566


git-svn-id: https://svn.freebsd.org/base/stable/11@338171 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f