MFC r338120: config(8): Allow escape-quoted empty strings
For use with things like BOOT_TAG=\"\" -- there are valid reasons to allow
empty strings, especially as these are usually being passed through as
options. The same argument could perhaps be made for the unquoted
variant in things like MODULES_OVERRIDE="", but it's not immediately clear
that this is an issue so I've left it untouched.


git-svn-id: https://svn.freebsd.org/base/stable/11@338227 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f