MFC r309748 (by glebius):
Treat R_X86_64_PLT32 relocs as R_X86_64_PC32.

If we load a binary that is designed to be a library, it produces
relocatable code via assembler directives in the assembly itself
(rather than compiler options).  This emits R_X86_64_PLT32 relocations,
which are not handled by the kernel linker.

Submitted by:	gallatin
Reviewed by:	kib
PR:		231451


git-svn-id: https://svn.freebsd.org/base/stable/9@338756 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f