MFC r336587:
tftpd(8): when completing an WRQ, flush the file before acknowleding receipt

tftpd(8) should flush a newly written file to disk before ACKing the final DATA
packet.  Otherwise there is a narrow race window when a subsequent read may not
see the file.  This is somewhat related to r330710, but the race window is much
smaller.  Hopefully this will fix the intermittent tests in Jenkins.

Reported by:	Jenkins


git-svn-id: https://svn.freebsd.org/base/stable/11@339049 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f