MFC r336594:
Fix tmpfs detection in the sys/fs/tmpfs tests

This code was originally written for NetBSD.  r306031 tried to adapt it to
FreeBSD, but didn't correctly handle the case that tmpfs was available, but
not already loaded.  Fix the logic to load the module if necessary.  The
tmpfs tests shouldn't be skipped anymore.

Also, fix a comment that was dislocated by r306031.

Reported by:	Jenkins


git-svn-id: https://svn.freebsd.org/base/stable/11@339050 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f