MFC r338216:
tftpd: Fix data corruption bug with netascii

Transferring files in netascii format requires, among other things,
translating all CR characters to a CR,NUL pair. tftpd does this correctly
except when the CR occurs as the last octet of a packet. In that case, it
erroneously drops the NUL which should be part of the following packet. The
bug was caused by using 0 as a sentinel value in a variable that could
legitimately hold 0. Fix it by switching the sentinel value to -1.

PR:		178055
Reported by:	Richard <rsitze@gmail.com>
Reviewed by:	cem
Differential Revision:	https://reviews.freebsd.org/D16853


git-svn-id: https://svn.freebsd.org/base/stable/11@339057 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f