MFC r338646: dd(1): Correct padding in status=progress
Output padding is specified via outlen, which is set using the return value
of fprintf. Because it's printing that padding plus a trailing byte, it
grows by one each iteration rather than reflecting actual length.

Additionally, iec was sized improperly for scaling up similarly to si.
Fixing this revealed that the humanize_number(3) call to populate persec
was using the wrong width.


git-svn-id: https://svn.freebsd.org/base/stable/11@339163 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f